% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trewjb.R
\name{iframeJbrowse}
\alias{iframeJbrowse}
\title{Create iframe view for Jbrowse from a url}
\usage{
iframeJbrowse(link, width = NULL, height = NULL, elementId = NULL)
}
\description{
A sample url is
"http://180.208.58.19/jbrowse/?data=data/hg19&loc=chr6:30309362..30310357&tracks=DNA,all_m6A,gene_model&highlight=chr6:30309513..30310230&nav=0&tracklist=0&overview=0" .
}
\examples{
iframeJbrowse("http://180.208.58.19/jbrowse/?data=data/hg19&loc=chr6:30309362..30310357&tracks=DNA,all_m6A,gene_model&highlight=chr6:30309513..30310230&nav=0&tracklist=0&overview=0")

}

