% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trewjb.R
\name{Jbrowse-shiny}
\alias{Jbrowse-shiny}
\alias{JbrowseOutput}
\alias{renderJbrowse}
\title{Shiny bindings for TREW Jbrowse iframe}
\usage{
JbrowseOutput(outputId, width = "100\%", height = "400px")

renderJbrowse(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a trewjb}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using trewjb within Shiny
applications and interactive Rmd documents.
}
\details{
\code{renderJbrowse} should return with \link[trewjb]{iframeJbrowse}.
}
\examples{
if (interactive() && require(shiny))
shinyApp(
    ui = fluidPage(
        fluidRow(
            JbrowseOutput(outputId = 'test')
        )
    ),
    server = function(input, output) {
        output$test <- renderJbrowse(
            iframeJbrowse('http://www.example.org')
        )
    }
)

}

