% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trewjb.R
\name{iframeJbrowse}
\alias{iframeJbrowse}
\title{Create iframe htmlwidget from a url.}
\usage{
iframeJbrowse(link, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{link}{Length-one character vector representing a url. It can be generagetd
by \link[trewjb]{getJbrowseLink}.}

\item{width, height, elementId}{Passed to htmlwidget.}
}
\description{
Create a htmlwidget that containing a iframe view for Jbrowse from a url.
}
\note{
It seems that the built-in RStudio viewer does not support iframe element,
    thus please consider using a web browser instead.
}
\examples{
if (interactive()) {
    url <- getJbrowseLink("hg19", "chr8", 92114847, 92231464,
        hight_start = 92150113, hight_end = 92152064,
        tracks = c("DNA", "gene_model", "all_m6A"),
        show_tracklist = TRUE, show_overview = TRUE)
    iframeJbrowse(url)
}

}

